move controller into dedicated package
