+++ Software/view/view.py
@@ -1,5 +1,3 @@
-import statistics
-from datetime import datetime
 from tkinter import Tk, Label
